Fix to #11687 - Query with projection cast on correlated collections throws on 2.1.0-preview2
Problem was that we were not recognizing a pattern where ToList<T> was used to cast a subquery returning collection of derived entities into their base types. We were trying to use correlated collection optimization pattern here but the logic was not accounting for the discrepancy in types.

Fix is to recognize the pattern and inject Cast<T>() into the subquery, which prevents correlated collection optimization from kicking in.