Metadata: Fix for calling OwnsOne on navigation which implements ICollection throws NullRef
Issue: Conventions create one-to-many relationship when navigation of sequence type is found.
Later when OwnsOne is called, we would try to reuse the same navigation's FK to convert it to ownership.
This works fine for reference navigation. For navigation implementing ICollection, we would have FK with different target type. Since we cannot do such conversion, we encounter exception in pipeline.

Fix is to remove FK if the target type is different from target type of FK.

This also uncovered another issue in query. If the owned entity has non-shadow PK, it will be used as FK too.
This does not work well with SelectExpression.PushDownSubquery() because we consider 2 column with same name as same in SQL tree and hence not alias it but still add it. Which generates invalid SQL.
As for fix in query, we would de-dup the inner projections in the subquery, since that does not cause issue with valuebuffer indexes in shapers. At outer level they would still be repeated.

Resolves #11688